services/webapp/requirements.txt
Django==2.2.1
psycopg2==2.8
pytz==2022.7.1
djangorestframework==3.9.3
django-rest-swagger==2.2.0
dateutils==0.6.6
uwsgi==2.0.20